refactor

+++ bob/sources.py
@@ -28,9 +28,7 @@ class Sources:
     def __exit__(self, type: Any, value: Any, tb: Any) -> None:
         return
 
-    def addSource(
-        self, coord: np.ndarray, sed: np.ndarray, time: Union[List[float], float] = None
-    ) -> None:
+    def addSource(self, coord: np.ndarray, sed: np.ndarray, time: Union[List[float], float] = None) -> None:
         if np.ndim(coord) == 1:
             coord, sed = [coord], [sed]
         time = np.zeros(len(coord)) if time is None else time
@@ -52,25 +50,15 @@ class Sources:
             self.nFreq = nFreq
             self.nSigma = nSigma
             self.nEnergy = nEnergy
-            self.sigma = (
-                np.fromfile(f, dtype="f8", count=nSigma) if nSigma > 0 else None
-            )
-            self.energy = (
-                np.fromfile(f, dtype="f8", count=nEnergy) if nEnergy > 0 else None
-            )
-            self.coord = np.fromfile(f, dtype="f8", count=nSources * 3).reshape(
-                (nSources, 3)
-            )
-            self.sed = np.fromfile(f, dtype="f8", count=nSources * nFreq).reshape(
-                (nSources, nFreq)
-            )
+            self.sigma = np.fromfile(f, dtype="f8", count=nSigma) if nSigma > 0 else None
+            self.energy = np.fromfile(f, dtype="f8", count=nEnergy) if nEnergy > 0 else None
+            self.coord = np.fromfile(f, dtype="f8", count=nSources * 3).reshape((nSources, 3))
+            self.sed = np.fromfile(f, dtype="f8", count=nSources * nFreq).reshape((nSources, nFreq))
             self.time = np.fromfile(f, dtype="f8", count=nSources)
 
     def write(self, fileName: Path) -> None:
         with open(fileName, "wb") as f:
-            np.array([self.nSigma, self.nEnergy, self.nSources, self.nFreq]).astype(
-                "u4"
-            ).tofile(f)
+            np.array([self.nSigma, self.nEnergy, self.nSources, self.nFreq]).astype("u4").tofile(f)
             if self.sigma is not None:
                 self.sigma.astype("f8").tofile(f)
             if self.energy is not None:
